[MIG] attachment_preview: Migration to 15.0
