+++ attachment_preview/i18n/vi_VN.po
@@ -6,8 +6,8 @@ msgid ""
 msgstr ""
 "Project-Id-Version: Odoo Server 15.0\n"
 "Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2022-08-23 14:25+0000\n"
-"PO-Revision-Date: 2022-08-23 14:25+0000\n"
+"POT-Creation-Date: 2022-08-23 14:29+0000\n"
+"PO-Revision-Date: 2022-08-23 14:29+0000\n"
 "Last-Translator: \n"
 "Language-Team: \n"
 "MIME-Version: 1.0\n"
@@ -18,21 +18,21 @@ msgstr ""
 #. module: attachment_preview
 #: model:ir.model,name:attachment_preview.model_ir_attachment
 msgid "Attachment"
-msgstr ""
+msgstr "Tập tin đính kèm"
 
 #. module: attachment_preview
 #. openerp-web
 #: code:addons/attachment_preview/static/src/xml/attachment_preview.xml:0
 #, python-format
 msgid "Open in new page"
-msgstr ""
+msgstr "Mở trong trang mới"
 
 #. module: attachment_preview
 #. openerp-web
 #: code:addons/attachment_preview/static/src/xml/attachment_preview.xml:0
 #, python-format
 msgid "Preview"
-msgstr ""
+msgstr "Xem trước"
 
 #. module: attachment_preview
 #. openerp-web
@@ -40,4 +40,4 @@ msgstr ""
 #: code:addons/attachment_preview/static/src/js/components/chatter/chatter.js:0
 #, python-format
 msgid "Preview %s"
-msgstr ""
+msgstr "Xem trước %s"
