+++ attachment_preview/static/src/scss/attachment_preview.scss
@@ -55,3 +55,10 @@
         }
     }
 }
+
+.o_group {
+    .mr-2,
+    .mx-2 {
+        margin-left: 0.5rem !important;
+    }
+}
